pyproject.toml
[tool.poetry]
name = "karmabot"
version = "0.1.0"
description = ""
authors = ["Brandon Richards <bqrichards@icloud.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
